LPS-83272 move files into api
