+++ modules/apps/currency-converter/currency-converter-api/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Currency Converter API
+Bundle-SymbolicName: com.liferay.currency.converter.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.currency.converter.model,\
+	com.liferay.currency.converter.util
