+++ modules/apps/currency-converter/currency-converter-web/src/main/java/com/liferay/currency/converter/web/internal/portlet/validator/CurrencyConverterPreferencesValidator.java
@@ -15,8 +15,8 @@
 package com.liferay.currency.converter.web.internal.portlet.validator;
 
 import com.liferay.currency.converter.web.internal.constants.CurrencyConverterPortletKeys;
-import com.liferay.currency.converter.web.internal.model.CurrencyConverter;
-import com.liferay.currency.converter.web.internal.util.CurrencyConverterUtil;
+import com.liferay.currency.converter.model.CurrencyConverter;
+import com.liferay.currency.converter.util.CurrencyConverterUtil;
 
 import java.util.ArrayList;
 import java.util.List;
