+++ modules/apps/currency-converter/currency-converter-web/src/main/resources/META-INF/resources/init.jsp
@@ -24,8 +24,8 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
 <%@ page import="com.liferay.currency.converter.web.configuration.CurrencyConverterConfiguration" %><%@
-page import="com.liferay.currency.converter.web.internal.model.CurrencyConverter" %><%@
-page import="com.liferay.currency.converter.web.internal.util.CurrencyConverterUtil" %><%@
+page import="com.liferay.currency.converter.model.CurrencyConverter" %><%@
+page import="com.liferay.currency.converter.util.CurrencyConverterUtil" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
 page import="com.liferay.portal.kernel.util.HtmlUtil" %><%@
